move out the attention processor.
